feat: Implement transaction management features
- Added TransactionNotFoundException for handling missing transactions.
- Updated Transaction model to use BigDecimal for amount and made description nullable.
- Introduced CreateTransactionProps and UpdateTransactionProps for transaction creation and updates.
- Created TransactionRepository interface and its implementation for database operations.
- Developed TransactionSqlEntity for mapping transactions to the database.
- Implemented TransactionChangeEventProducer for producing events related to transaction changes.
- Created REST endpoints for transaction management including create, update, retrieve, and delete operations.
- Added request and response data classes for handling transaction data in REST API.
- Implemented database migration script for creating the transactions table.
- Added unit tests for TransactionService and mock factories for testing transaction-related functionalities.

+++ src/main/kotlin/org/balhom/transactionsapi/modules/transactions/infrastructure/producers/data/TransactionChangeEvent.kt
@@ -1,14 +1,15 @@
 package org.balhom.transactionsapi.modules.transactions.infrastructure.producers.data
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class TransactionChangeEvent(
     val action: String,
     val id: UUID,
-    val amount: Double,
-    val oldAmount: Double?,
-    val cpGoalMonthlySaving: Double,
-    val cpGoalYearlySaving: Double,
+    val amount: BigDecimal,
+    val oldAmount: BigDecimal?,
+    val cpGoalMonthlySaving: BigDecimal,
+    val cpGoalYearlySaving: BigDecimal,
     val currencyProfileId: UUID,
     val userId: UUID,
 )
